modify RssFeedClass, separate common function to abstract class
