+++ app/src/main/java/me/rei_m/hbfavkotlin/models/NewEntryModel.kt
@@ -41,7 +41,7 @@ public class NewEntryModel {
             }
         }
 
-        NewEntryRss.request()
+        NewEntryRss().request()
                 .onBackpressureBuffer()
                 .subscribeOn(Schedulers.newThread())
                 .observeOn(AndroidSchedulers.mainThread())
