+++ app/src/main/java/me/rei_m/hbfavkotlin/network/AbsEntryRss.kt
@@ -1,0 +1,45 @@
+package me.rei_m.hbfavkotlin.network
+
+import com.squareup.okhttp.CacheControl
+import com.squareup.okhttp.HttpUrl
+import com.squareup.okhttp.OkHttpClient
+import com.squareup.okhttp.Request
+import kotlinx.dom.parseXml
+import me.rei_m.hbfavkotlin.entities.EntryEntity
+import me.rei_m.hbfavkotlin.utils.RssXmlUtils
+import rx.Observable
+import java.net.HttpURLConnection
+
+public abstract class AbsEntryRss {
+
+    protected fun request(url: HttpUrl): Observable<EntryEntity> {
+
+        return Observable.create({ t ->
+
+            val request = Request.Builder()
+                    .url(url)
+                    .cacheControl(CacheControl.FORCE_NETWORK)
+                    .build()
+
+            val response = OkHttpClient().newCall(request).execute()
+
+            if (response.code() == HttpURLConnection.HTTP_OK) {
+
+                val document = parseXml(response.body().byteStream())
+
+                val feeds = document.getElementsByTagName("item")
+
+                val feedCount = feeds.length
+
+                for (i_feed in 0..feedCount - 1) {
+                    t.onNext(RssXmlUtils.createEntryFromFeed(feeds.item(i_feed)))
+                }
+
+            } else {
+                t.onError(Throwable(response.code().toString()))
+            }
+
+            t.onCompleted()
+        })
+    }
+}
