+++ app/src/main/java/me/rei_m/hbfavkotlin/network/NewEntryRss.kt
@@ -1,55 +1,20 @@
 package me.rei_m.hbfavkotlin.network
 
-import com.squareup.okhttp.CacheControl
 import com.squareup.okhttp.HttpUrl
-import com.squareup.okhttp.OkHttpClient
-import com.squareup.okhttp.Request
-import kotlinx.dom.parseXml
 import me.rei_m.hbfavkotlin.entities.EntryEntity
-import me.rei_m.hbfavkotlin.utils.RssXmlUtils
 import rx.Observable
-import java.net.HttpURLConnection
 
-public final class NewEntryRss private constructor() {
+public final class NewEntryRss : AbsEntryRss() {
 
-    companion object {
+    public fun request(): Observable<EntryEntity> {
 
-        public fun request(): Observable<EntryEntity> {
+        val url = HttpUrl.Builder()
+                .scheme("http")
+                .host("b.hatena.ne.jp")
+                .addPathSegment("entrylist")
+                .addQueryParameter("mode", "rss")
+                .build()
 
-            return Observable.create({ t ->
-
-                val url = HttpUrl.Builder()
-                        .scheme("http")
-                        .host("b.hatena.ne.jp")
-                        .addPathSegment("entrylist")
-                        .addQueryParameter("mode", "rss")
-                        .build()
-
-                val request = Request.Builder()
-                        .url(url)
-                        .cacheControl(CacheControl.FORCE_NETWORK)
-                        .build()
-
-                val response = OkHttpClient().newCall(request).execute()
-
-                if (response.code() == HttpURLConnection.HTTP_OK) {
-
-                    val document = parseXml(response.body().byteStream())
-
-                    val feeds = document.getElementsByTagName("item")
-
-                    val feedCount = feeds.length
-
-                    for (i_feed in 0..feedCount - 1) {
-                        t.onNext(RssXmlUtils.createEntryFromFeed(feeds.item(i_feed)))
-                    }
-
-                } else {
-                    t.onError(Throwable(response.code().toString()))
-                }
-
-                t.onCompleted()
-            })
-        }
+        return super.request(url)
     }
 }
